# GH-4224 SHACL - Support sequence paths and alternate paths