--- a/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ShaclSail.java
+++ b/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ShaclSail.java
@@ -176,7 +176,8 @@ public class ShaclSail extends ShaclSailBaseConfiguration {
 
 	// lockManager used for read/write locks used to synchronize validation so that SNAPSHOT isolation is sufficient to
 	// achieve SERIALIZABLE isolation wrt. validation
-	final ReadPrefReadWriteLockManager serializableValidationLock = new ReadPrefReadWriteLockManager();
+	final ReadPrefReadWriteLockManager serializableValidationLock = new ReadPrefReadWriteLockManager(
+			"ShaclSail_SerializableValidation");
 
 	// shapesCacheLockManager used to keep track of changes to the cache
 	private StampedLockManager.Cache<List<ContextWithShapes>> cachedShapes;
@@ -311,6 +312,7 @@ public class ShaclSail extends ShaclSailBaseConfiguration {
 				SHACL.HAS_VALUE,
 				SHACL.TARGET_PROP,
 				SHACL.INVERSE_PATH,
+				SHACL.ALTERNATIVE_PATH,
 				SHACL.NODE,
 				SHACL.QUALIFIED_MAX_COUNT,
 				SHACL.QUALIFIED_MIN_COUNT,