--- a/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/ShaclProperties.java
+++ b/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/ShaclProperties.java
@@ -108,16 +108,17 @@ public class ShaclProperties {
 
 				switch (predicate) {
 				case "http://www.w3.org/1999/02/22-rdf-syntax-ns#type":
-					if (object.stringValue().equals("http://www.w3.org/ns/shacl#NodeShape")) {
+					if (object.equals(SHACL.NODE_SHAPE)) {
 						if (type != null && !type.equals(SHACL.NODE_SHAPE)) {
 							throw new IllegalStateException(
-									"Shape with multiple types: <" + type + ">, <" + SHACL.NODE_SHAPE + ">");
+									"Shape " + id + " with multiple types: <" + type + ">, <" + SHACL.NODE_SHAPE + ">");
 						}
 						type = SHACL.NODE_SHAPE;
-					} else if (object.stringValue().equals("http://www.w3.org/ns/shacl#PropertyShape")) {
+					} else if (object.equals(SHACL.PROPERTY_SHAPE)) {
 						if (type != null && !type.equals(SHACL.PROPERTY_SHAPE)) {
 							throw new IllegalStateException(
-									"Shape with multiple types: <" + type + ">, <" + SHACL.PROPERTY_SHAPE + ">");
+									"Shape " + id + " with multiple types: <" + type + ">, <" + SHACL.PROPERTY_SHAPE
+											+ ">");
 						}
 						type = SHACL.PROPERTY_SHAPE;
 					}
@@ -252,7 +253,12 @@ public class ShaclProperties {
 					if (path != null) {
 						throw new IllegalStateException(predicate + " already populated");
 					}
-					assert type != SHACL.NODE_SHAPE;
+					if (type == null) {
+						type = SHACL.PROPERTY_SHAPE;
+					} else if (!type.equals(SHACL.PROPERTY_SHAPE)) {
+						throw new IllegalStateException("Shape " + id
+								+ " has sh:path and must be of type sh:PropertyShape but is type " + type);
+					}
 					path = (Resource) object;
 					break;
 				case "http://www.w3.org/ns/shacl#in":
@@ -323,7 +329,7 @@ public class ShaclProperties {
 
 		// We default to sh:NodeShape if no other type is given.
 		if (type == null) {
-			type = SHACL.NODE_SHAPE;
+			type = path == null ? SHACL.NODE_SHAPE : SHACL.PROPERTY_SHAPE;
 		}
 
 	}