--- a/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/Targetable.java
+++ b/core/sail/shacl/src/main/java/org/eclipse/rdf4j/sail/shacl/ast/Targetable.java
@@ -10,17 +10,14 @@
  *******************************************************************************/
 package org.eclipse.rdf4j.sail.shacl.ast;
 
-import java.util.stream.Stream;
+import java.util.Set;
 
 import org.eclipse.rdf4j.sail.shacl.wrapper.data.RdfsSubClassOfReasoner;
 
 public interface Targetable {
 
-	Stream<StatementMatcher> getStatementMatcher(StatementMatcher.Variable subject, StatementMatcher.Variable object,
-			RdfsSubClassOfReasoner rdfsSubClassOfReasoner);
-
-	String getTargetQueryFragment(StatementMatcher.Variable subject, StatementMatcher.Variable object,
+	SparqlFragment getTargetQueryFragment(StatementMatcher.Variable subject, StatementMatcher.Variable object,
 			RdfsSubClassOfReasoner rdfsSubClassOfReasoner,
-			StatementMatcher.StableRandomVariableProvider stableRandomVariableProvider);
+			StatementMatcher.StableRandomVariableProvider stableRandomVariableProvider, Set<String> inheritedVarNames);
 
 }