--- a/core/sail/shacl/src/test/resources/test-cases/class/and2/invalid/case9/query1.rq
+++ b/core/sail/shacl/src/test/resources/test-cases/class/and2/invalid/case9/query1.rq
@@ -12,4 +12,6 @@ ex:validPerson1 a ex:SecondTarget ;
 ex:peter a ex:SecondTarget.
 ex:steve a ex:SecondTarget.
 
+ex:peter a ex:Person, ex:Human.
+
 }