@hmottestad 2023-01-15
TODO

- [ ] Run benchmarks
- [ ] Create new benchmarks
- [ ] Split up the scope of the task so that we can merge the currently supported paths