@hmottestad 2023-02-09
## Before
```
```

## After
```
Benchmark                                                                      (NUMBER_OF_TRANSACTIONS)  (existingTargets)                         (shape)  Mode  Cnt    Score   Error  Units
AddRemoveBenchmarkEmpty.shacl                                                                       N/A                N/A                             N/A  avgt    2   16.348          ms/op
ClassBenchmarkEmpty.shacl                                                                           N/A                N/A                             N/A  avgt    2   26.744          ms/op
ClassBenchmarkEmpty.shaclBulk                                                                       N/A                N/A                             N/A  avgt    2   24.780          ms/op
ComplexBenchmark.shacl                                                                              N/A                N/A                             N/A  avgt    2    5.457          ms/op
ComplexBenchmark.shaclCache                                                                         N/A                N/A                             N/A  avgt    2    5.654          ms/op
ComplexBenchmark.shaclEmptyTransactions                                                             N/A                N/A                             N/A  avgt    2    1.506          ms/op
ComplexBenchmark.shaclNoTransactions                                                                N/A                N/A                             N/A  avgt    2    1.450          ms/op
ComplexBenchmark.shaclNothingToValidateTransactions                                                 N/A                N/A                             N/A  avgt    2    2.623          ms/op
ComplexBenchmark.shaclParallel                                                                      N/A                N/A                             N/A  avgt    2    5.398          ms/op
ComplexBenchmark.shaclParallelCache                                                                 N/A                N/A                             N/A  avgt    2    5.418          ms/op
ComplexBenchmark.shaclParallelCacheSingleTransactionNoIsolation                                     N/A                N/A                             N/A  avgt    2    4.734          ms/op
ComplexBenchmark.shaclPropertiesSwitch                                                              N/A                N/A                             N/A  avgt    2    0.055          ms/op
ComplexTargetBenchmark.benchmark                                                                    N/A                  1                             N/A  avgt    2   29.140          ms/op
ComplexTargetBenchmark.benchmark                                                                    N/A               1000                             N/A  avgt    2   14.504          ms/op
ComplexTargetBenchmark.benchmark                                                                    N/A             100000                             N/A  avgt    2   13.340          ms/op
DatatypeBenchmarkEmpty.shacl                                                                        N/A                N/A                             N/A  avgt    2   10.573          ms/op
DatatypeBenchmarkEmpty.shaclBulk                                                                    N/A                N/A                             N/A  avgt    2   12.993          ms/op
DatatypeBenchmarkLinear.shacl                                                                         1                N/A                             N/A  avgt    2    0.360          ms/op
DatatypeBenchmarkLinear.shacl                                                                        10                N/A                             N/A  avgt    2    3.220          ms/op
DatatypeBenchmarkLinear.shacl                                                                       100                N/A                             N/A  avgt    2   31.681          ms/op
DatatypeBenchmarkPrefilled.shacl                                                                    N/A                N/A                             N/A  avgt    2   16.687          ms/op
DatatypeBenchmarkSerializableEmpty.shacl                                                            N/A                N/A                             N/A  avgt    2   22.910          ms/op
HasValueBenchmarkEmpty.shacl                                                                        N/A                N/A                             N/A  avgt    2   33.034          ms/op
HasValueBenchmarkEmpty.shaclBulk                                                                    N/A                N/A                             N/A  avgt    2   32.223          ms/op
LanguageInBenchmarkEmpty.shacl                                                                      N/A                N/A                             N/A  avgt    2   21.765          ms/op
LanguageInBenchmarkEmpty.shaclBulk                                                                  N/A                N/A                             N/A  avgt    2   32.794          ms/op
MaxCountBenchmarkEmpty.shacl                                                                        N/A                N/A                             N/A  avgt    2   10.945          ms/op
MaxCountBenchmarkEmpty.shaclBulk                                                                    N/A                N/A                             N/A  avgt    2   14.077          ms/op
MinCountBenchmarkEmpty.shacl                                                                        N/A                N/A                             N/A  avgt    2    8.572          ms/op
MinCountBenchmarkEmpty.shaclBulk                                                                    N/A                N/A                             N/A  avgt    2   12.831          ms/op
MinCountBenchmarkEmpty.shaclClear                                                                   N/A                N/A                             N/A  avgt    2   19.457          ms/op
MinCountBenchmarkEmpty.shaclMinCountZero                                                            N/A                N/A                             N/A  avgt    2    8.115          ms/op
MinCountBenchmarkPrefilled.shacl                                                                    N/A                N/A                             N/A  avgt    2   14.261          ms/op
MinCountPrefilledVsEmptyBenchmark.shaclEmpty                                                        N/A                N/A                             N/A  avgt    2   32.464          ms/op
MinCountPrefilledVsEmptyBenchmark.shaclEmptyJustInitializeAndEmptyTransaction                       N/A                N/A                             N/A  avgt    2    7.349          ms/op
MinCountPrefilledVsEmptyBenchmark.shaclEmptyJustinit                                                N/A                N/A                             N/A  avgt    2    3.007          ms/op
MinCountPrefilledVsEmptyBenchmark.shaclPrefilled                                                    N/A                N/A                             N/A  avgt    2   31.291          ms/op
NotClassBenchmarkEmpty.shacl                                                                        N/A                N/A                             N/A  avgt    2   67.441          ms/op
NotClassBenchmarkEmpty.shaclBulk                                                                    N/A                N/A                             N/A  avgt    2   56.811          ms/op
NotClassBenchmarkEmpty.shaclWithoutAnimals                                                          N/A                N/A                             N/A  avgt    2   35.109          ms/op
NotMaxCountBenchmarkEmpty.shacl                                                                     N/A                N/A                             N/A  avgt    2   24.686          ms/op
NotMaxCountBenchmarkEmpty.shaclBulk                                                                 N/A                N/A                             N/A  avgt    2   16.838          ms/op
NotUniqueLangBenchmarkEmpty.shacl                                                                   N/A                N/A                             N/A  avgt    2   91.948          ms/op
NotUniqueLangBenchmarkEmpty.shaclBulk                                                               N/A                N/A                             N/A  avgt    2   60.712          ms/op
OrDatatypeBenchmark.shacl                                                                           N/A                N/A                             N/A  avgt    2  135.479          ms/op
OrDatatypeBenchmark.shaclBulk                                                                       N/A                N/A                             N/A  avgt    2   60.619          ms/op
QualifiedValueShapeBenchmarkEmpty.shacl                                                             N/A                N/A                             N/A  avgt    2  713.014          ms/op
QualifiedValueShapeBenchmarkEmpty.shaclBulk                                                         N/A                N/A                             N/A  avgt    2  254.818          ms/op
RdfsReasonerBenchmarkEmpty.shacl                                                                    N/A                N/A                             N/A  avgt    2   44.213          ms/op
RdfsReasonerBenchmarkEmpty.shaclBulk                                                                N/A                N/A                             N/A  avgt    2   41.545          ms/op
ShaclLoadingBenchmark.loadComplexShapes                                                             N/A                N/A                             N/A  avgt    2    1.621          ms/op
ShaclLoadingBenchmark.testAddingDataAndAddingShapes                                                 N/A                N/A                             N/A  avgt    2   16.603          ms/op
TargetBenchmarkInitialData.benchmark                                                                N/A                  1  shaclDatatypeSparqlTarget.trig  avgt    2    8.841          ms/op
TargetBenchmarkInitialData.benchmark                                                                N/A                  1  shaclDatatypeTargetFilter.trig  avgt    2   25.358          ms/op
TargetBenchmarkInitialData.benchmark                                                                N/A               1000  shaclDatatypeSparqlTarget.trig  avgt    2    8.768          ms/op
TargetBenchmarkInitialData.benchmark                                                                N/A               1000  shaclDatatypeTargetFilter.trig  avgt    2   24.356          ms/op
TargetBenchmarkInitialData.benchmark                                                                N/A             100000  shaclDatatypeSparqlTarget.trig  avgt    2    2.728          ms/op
TargetBenchmarkInitialData.benchmark                                                                N/A             100000  shaclDatatypeTargetFilter.trig  avgt    2    8.898          ms/op
TargetShapeBenchmark.shacl                                                                          N/A                N/A                             N/A  avgt    2   17.722          ms/op
UniqueLangBenchmarkEmpty.shacBulk                                                                   N/A                N/A                             N/A  avgt    2   56.237          ms/op
UniqueLangBenchmarkEmpty.shacl                                                                      N/A                N/A                             N/A  avgt    2   36.908          ms/op
ValidationTupleBenchmark.randomData                                                                 N/A                N/A                             N/A  avgt    2  147.897          ms/op
ValueInBenchmarkEmpty.shacl                                                                         N/A                N/A                             N/A  avgt    2   33.255          ms/op
ValueInBenchmarkEmpty.shaclBulk                                                                     N/A                N/A                             N/A  avgt    2   38.847          ms/op
```